pyproject.toml
[tool.poetry]
name = "ai-hub"
version = "0.1.0"
description = ""
authors = ["Tezan Sahu <tezansahu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
streamlit = "^1.41.1"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
plantweb = "^1.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
